+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY7C3CEQY64RZEZ4HKWQTTMF"
+        "01KY9CSN4G5MVQNG2ZPQB0WZW9"
     )
     r"""The unique identifier of the data source"""