+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -250,9 +250,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-07-23T11:34:25.074Z"
+    created: Optional[str] = "2026-07-24T06:25:07.444Z"
 
-    updated: Optional[str] = "2026-07-23T11:34:25.074Z"
+    updated: Optional[str] = "2026-07-24T06:25:07.444Z"
 
     updated_by_id: OptionalNullable[str] = UNSET