+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY7C3CEJAHMWMGDCVD7G83WK"
+        "01KY9CSN4BMPWJKBP7DFD339MB"
     )
     r"""The unique identifier of the data source"""