+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY7C3CEPNCGPR0CEQV1DY2H4"
+        "01KY9CSN4EW204G257KZ0GG0KA"
     )
     r"""The unique identifier of the data source"""