+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY7C3CEV9KTPBPXW0DGYSKND"
+        "01KY9CSN4KXNN4HBCNFB5E6P60"
     )
     r"""The unique identifier of the data source"""