+++ packages/orq-rc/RELEASES.md
@@ -9709,3 +9709,13 @@ Based on:
 - [python v4.13.0-rc.2] packages/orq-rc
 ### Releases
 - [PyPI v4.13.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.2 - packages/orq-rc
+
+## 2026-07-24 06:27:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.13.0-rc.3] packages/orq-rc
+### Releases
+- [PyPI v4.13.0-rc.3] https://pypi.org/project/orq-ai-sdk/4.13.0-rc.3 - packages/orq-rc